refactor: address CodeRabbit review findings in E2E tests
- Rename phase_01_faucet to phase_01_wallet_ui to match what the phase
  actually does (no faucet interaction; it verifies balance display and
  wallet UI elements)
- Run validation tests before SPV readiness gate in phase_05 since they
  are pure client-side checks that don't need SPV
- Assert action button is present in sub-test B rather than silently
  skipping the click path when absent
