+++ tests/e2e/phases/mod.rs
@@ -1,5 +1,5 @@
 pub mod phase_00_setup;
-pub mod phase_01_faucet;
+pub mod phase_01_wallet_ui;
 pub mod phase_02_wallet;
 pub mod phase_03_platform;
 pub mod phase_04_tokens;
