+++ tests/e2e/phases/phase_01_wallet_ui.rs
